requirements.txt
PySide6>=6.6
PyMuPDF>=1.23
python-docx>=1.1
ebooklib>=0.18
beautifulsoup4>=4.12
